+++ gnomad_bigquery/bq_import_gnomad.py
@@ -1,33 +1,46 @@
 import argparse
 import sys
-from gnomad_bigquery.v2 import bq_import
-from gnomad_bigquery.v2.bq_table_descriptions import get_variants_table_desc, get_meta_table_desc, get_genotypes_table_desc
-from gnomad_bigquery.v2.bq_utils import logger
+
+from gnomad_bigquery import bq_import
+from gnomad_bigquery.bq_table_descriptions import (get_genotypes_table_desc,
+                                                   get_meta_table_desc,
+                                                   get_variants_table_desc)
+from gnomad_bigquery.bq_utils import logger
 
 
 def main(args):
 
-    def import_data(data_type: str, data: str, description: str):
-        parquet_files = f"{args.input_dir}/gnomad_{data_type}_{data}.parquet/*.parquet"
-        logger.info(f"Importing {data_type} {data} from {parquet_files}")
+    def import_data(input_dir: str, version: int, data_type: str, data: str, description: str):
+        parquet_files = f"{input_dir}_{data}.parquet/*.parquet"
+        logger.info(f"Importing gnomAD v{version} {data_type} {data} from {parquet_files}")
         parser = bq_import.get_parser()
         imp_args = parser.parse_args(['--dataset', args.dataset,
                         '--parquet_files', parquet_files,
                         '--table', f'{data_type}_{data}',
-                        '--write_disposition', 'WRITE_TRUNCATE' if args.overwrite else 'WRITE_EMPTY',
+                        '--write_disposition', 
+                        'WRITE_TRUNCATE' if args.overwrite else 'WRITE_EMPTY',
                         '--description', description])
         bq_import.main(imp_args)
 
-    data_types = (['exomes'] if args.exomes else []) + (['genomes'] if args.genomes else [])
+    data_types = []
+    if args.exomes:
+        data_types.append("exomes")
+    if args.genomes:
+        data_types.append("genomes")
+
+    version = args.version
+
     for data_type in data_types:
+        input_dir = f"{args.input_dir}/gnomad_v{version}_{data_type}"
+
         if args.import_meta:
-            import_data(data_type, 'meta', get_meta_table_desc(data_type))
+            import_data(input_dir, version, data_type, 'meta', get_meta_table_desc(version, data_type))
 
         if args.import_variants:
-            import_data(data_type, 'variants', get_variants_table_desc(data_type))
+            import_data(input_dir, version, data_type, 'variants', get_variants_table_desc(version, data_type))
 
         if args.import_genotypes:
-            import_data(data_type, 'genotypes', get_genotypes_table_desc(data_type))
+            import_data(input_dir, version, data_type, 'genotypes', get_genotypes_table_desc(version, data_type))
 
 
 if __name__ == '__main__':
@@ -35,6 +48,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--exomes', help='Will import exomes data. At least one of --exomes or --genomes is required.', action='store_true')
     parser.add_argument('--genomes', help='Will import genomes data. At least one of --exomes or --genomes is required.', action='store_true')
+    parser.add_argument('--version', help='Version of gnomAD to import.', choices=[2,3], type=int, required=True)
     parser.add_argument('--dataset', help='Dataset to create the table in. (default: gnomad)', default='gnomad')
     parser.add_argument('--import_meta', help='Imports samples metadata.', action='store_true')
     parser.add_argument('--import_variants', help='Imports variants.', action='store_true')
@@ -45,5 +59,11 @@ if __name__ == '__main__':
 
     if not args.exomes and not args.genomes:
         sys.exit("At least one of --exomes or --genomes needs to be specified.")
+    
+    if args.exomes and args.version==3:
+        sys.exit('gnomAD v3 does not contain exomes')
 
+    if not args.import_genotypes and not args.import_variants and not args.import_metadata:
+        sys.exit('Error: At least one of --import_metadata, --import_variants or --import_genotypes must be specified')
+        
     main(args)
