+++ gnomad_bigquery/bq_table_descriptions.py
@@ -1,27 +1,28 @@
-from gnomad_hail import CSQ_ORDER
+from gnomad.utils.vep import CSQ_ORDER
 
 LEAST_CSQ = '3_prime_UTR_variant'
-MAX_AF = 0.02
 
-def get_variants_table_desc(data_type: str = None):
+def get_variants_table_desc(version: int, data_type: str = None):
     dtype = '' if data_type is None else f' {data_type}'
-    return f"""This table contains all gnomad{dtype} variants.
+    return f"""This table contains all gnomad v{version} {dtype} variants.
     Notes:
     * All variants were split, so multi-allelic sites are represented as multiple rows (one per non-ref allele)
     * Only VEP transcript annotations are available (non-coding annotations aren't)
     """
 
-def get_meta_table_desc(data_type: str = None):
+
+def get_meta_table_desc(version: int, data_type: str = None):
     dtype = '' if data_type is None else f' {data_type}'
-    return f"""This table contains all gnomad{dtype} samples metadata.
+
+    return f"""This table contains all gnomad v{version} {dtype} samples metadata.
     """
 
-def get_genotypes_table_desc(data_type: str  = None):
+
+def get_genotypes_table_desc(version: int, data_type: str = None):
     dtype = '' if data_type is None else f' {data_type}'
     all_str = 'all' if data_type is None else data_type
-    return f"""This table contains all gnomad{dtype} non-reference genotypes for variants meeting the following criteria:
-    1. Global variant allele frequency <= {int(100*MAX_AF)}%
-    2. Variant VEP annotation is at least `{LEAST_CSQ}` (i.e. one of: {','.join(CSQ_ORDER[0:CSQ_ORDER.index(LEAST_CSQ) + 1])})
+    return f"""This table contains all gnomad v{version} {dtype} non-reference genotypes for variants meeting the following criteria:
+    1. Variant VEP annotation is at least `{LEAST_CSQ}` (i.e. one of: {','.join(CSQ_ORDER[0:CSQ_ORDER.index(LEAST_CSQ) + 1])})
     
     A few notes about data representation:
     * All variants were split beforehand, so all genotypes are bi-allelic.
@@ -32,7 +33,7 @@ def get_genotypes_table_desc(data_type: str  = None):
     """
 
 
-def get_data_view_desc(data_type: str = None):
+def get_data_view_desc(version: int, data_type: str = None):
     if data_type is not None:
         view_tables_str = f'{data_type}_variants, {data_type}_genotypes and {data_type}_meta tables'
         n_tables_str = '3 tables.'
@@ -40,5 +41,5 @@ def get_data_view_desc(data_type: str = None):
         view_tables_str = 'exomes and genomes views.'
         n_tables_str = '2 views'
 
-    return f"""This view contains a flat version of joining the {view_tables_str}.
+    return f"""This view contains a flat version of joining the gnomAD v{version} {view_tables_str}.
     To get more information about the content of this view, take a look a the description of those {n_tables_str}."""
