--- gnomad_bigquery/v3/__init__.py
@@ -1,2 +1,0 @@
-from .bq_utils import *
-from .bq_table_descriptions import *
